requirements.txt
Django~=2.2.10
django-widget-tweaks~=1.4.8
django-allauth~=0.41.0
stripe~=2.27.0
Pillow~=2.2.1
